stage 1 complete

+++ app/app.py
@@ -13,12 +13,10 @@ from kivy.core.window import Window
 from kivy.utils import get_color_from_hex
 from kivy.graphics import Color, Rectangle
 
-
 Builder.load_file('design.kv')
 
 
 BACKGROUND_COLOR = '#ffffff'
-# BACKGROUND_COLOR = '#636661'
 
 class CustomErrorPopup(Popup):
 	pass
@@ -48,36 +46,52 @@ class CustomAddPopup(Popup):
 		self.dirname = self.ids._chapter.text.strip().lower()
 		self.filename = self.ids._topic.text.strip().lower()
 		self.info = self.ids._info.text.strip()
-		pat = re.compile(".{1,10}$")
+		pat = re.compile("^[a-z0-9]{1,15}$")
+
+		#validating input data
 		if (not pat.match(self.dirname) or not pat.match(self.filename)):
-			self.ids.no_validate.text = "Invalid input for \n Chapter or Topic"
+			self.ids.no_validate.text = "Invalid input for \nChapter or Topic"
 			return
 		elif (not self.info):
-			self.ids.no_validate.text = "Info field cannot be empty!!"
+			self.ids.no_validate.text = "Content field cannot be empty!!"
 			return
+
+		#if validated, enable add button
 		self.ids.no_validate.text = ""
 		self.ids.add_button.disabled= False
 
 
 	def _add(self):
 		self.dismiss()
+
+		# set the text fields of the popup to empty  
 		self.ids._chapter.text = ""
 		self.ids._topic.text = ""
 		self.ids._info.text = ""
+
+		#disable add button once data is added
 		self.ids.add_button.disabled= True
 		self.data_storage()
 
 
 	def data_storage(self):
 
-		#list new directory name into dirs list
+		#lists new directory(chapter) name into dirs list
 		dir_exst = False 
-		with open("../data/dirs.txt","r+") as fp:
+		with open("../data/dirs_list.txt","r+") as fp:
+
+			#check if the directory is already present
 			for dirt in fp.readlines():
 				if dirt[:-1] == self.dirname:
 					dir_exst = True
+			
+			# if not, create a dir and update dirs list
 			if not dir_exst:
+
+				# create directory
 				os.system(f"mkdir ../data/{self.dirname}")
+
+				#updating dir list and add a button to the main object
 				fp.write(self.dirname+'\n')
 				btn = RoundedButton(
 					text = self.dirname.upper(),
@@ -88,22 +102,30 @@ class CustomAddPopup(Popup):
 				self.main_obj.ids.chapter_.add_widget(btn, index =1)
 		fp.close()
 		
-		# list new filename into files list
+		# lists new filename(topic) into files list and creates it if dosen't exist
 		file_exst = False
 		try:
-			with open(f"../data/{self.dirname}/files.txt","r+") as fp:
+			with open(f"../data/{self.dirname}/files_list.txt","r+") as fp:
+
+				# checks if file already listed in files
 				for file in fp.readlines():
 					if file[:-1] == self.filename:
 						file_exst = True
+
+				#write if not
 				if not file_exst:
 					fp.write(self.filename+'\n')
 			fp.close()
+
 		except FileNotFoundError:
-			with open(f"../data/{self.dirname}/files.txt","w") as fp:
+
+			#creates a new file and update it
+			with open(f"../data/{self.dirname}/files_list.txt","w") as fp:
 				fp.write(self.filename+'\n')
 			fp.close()
 
-		# Create a new file and add data to it
+
+		# Update the topic file, creates if dosen't exist
 		with open(f"../data/{self.dirname}/{self.filename}.txt","a") as fp:
 			fp.write(self.info+'\n')
 		fp.close()
@@ -111,6 +133,7 @@ class CustomAddPopup(Popup):
 
 
 class CustomRemovePopup(Popup):
+
 	def __init__(self, **kwargs):
 		super(CustomRemovePopup, self).__init__(**kwargs)
 
@@ -125,20 +148,29 @@ class CustomRemovePopup(Popup):
 		self.ids._chapter_.text = ""
 		self.ids._topic_.text = ""
 
+		# if dir or filename is not given 
 		if (not dirname or not filename):
 			error_pop.ids.err.text = "Input fields cannot be empty"
 			error_pop.open()
 
+
+		# deletes the whole dir(chapter)
 		if (filename == "*" or filename == "all"):
-			dir_exst = False 
-			with open("../data/dirs.txt","r+") as fp:
+			dir_exst = False
+
+			with open("../data/dirs_list.txt","r+") as fp:
+				
+				#look for dir in dirs lists
 				for dirt in fp.readlines():
 					if dirt[:-1] == dirname:
 						dir_exst = True
+
+				#alert if non-existant
 				if not dir_exst:
 					error_pop.ids.err.text = "Chapter Not Found"
 					error_pop.open()
 				else:
+					# removes if found
 					fp.seek(0,0)
 					ft = open("../data/dirs2.txt", "w")
 					for dirt in fp.readlines():
@@ -146,21 +178,31 @@ class CustomRemovePopup(Popup):
 							continue
 						ft.write(dirt)
 					ft.close()
-					os.remove("../data/dirs.txt")
-					os.rename("../data/dirs2.txt", "../data/dirs.txt")
-					os.system(f"rm -rf ../data/{dirname}")
 			fp.close()
+
+			if dir_exst:
+				# removes the whole chapter if found
+				os.remove("../data/dirs_list.txt")
+				os.rename("../data/dirs2.txt", "../data/dirs_list.txt")
+				os.system(f"rm -rf ../data/{dirname}")
+
 		
+		# removes a topic
 		else:
 			file_exst = False
 			try:
-				with open(f"../data/{dirname}/files.txt","r+") as fp:
+				with open(f"../data/{dirname}/files_list.txt","r+") as fp:
+					
+					#look for topic
 					for file in fp.readlines():
 						if file[:-1] == filename:
 							file_exst = True
+
+					#raises an error if non existant
 					if not file_exst:
 						error_pop.ids.err.text = "Topic Not Found"
 						error_pop.open()
+
 					else:
 						fp.seek(0,0)
 						ft = open(f"../data/{dirname}/files2.txt","w")
@@ -169,15 +211,19 @@ class CustomRemovePopup(Popup):
 								continue
 							ft.write(file)
 						ft.close()
-						os.remove(f"../data/{dirname}/files.txt")
-						os.rename(f"../data/{dirname}/files2.txt", f"../data/{dirname}/files.txt")
-						os.remove(f"../data/{dirname}/{filename}.txt")
 				fp.close()
+				if file_exst:
+					os.remove(f"../data/{dirname}/files_list.txt")
+					os.rename(f"../data/{dirname}/files2.txt", f"../data/{dirname}/files_list.txt")
+					os.remove(f"../data/{dirname}/{filename}.txt")
+
 			except:
 				error_pop.ids.err.text = "Chapter Not Found"
 				error_pop.open()
 
 		self.dismiss()
+
+		#updates the main object
 		self.main_obj.add_chapters()
 
 
@@ -186,6 +232,8 @@ class MyLayout(Widget):
 
 	def __init__(self,**kwargs):
 		super(MyLayout, self).__init__(**kwargs)
+
+		#add chapters to the screen
 		self.add_chapters()
 		self.add_popup = CustomAddPopup()
 		self.remove_popup = CustomRemovePopup()
@@ -209,7 +257,7 @@ class MyLayout(Widget):
 		)
 		self.ids.chapter_.add_widget(rmv_btn)
 		try:
-			with open("../data/dirs.txt",'r') as fp:
+			with open("../data/dirs_list.txt",'r') as fp:
 				for line in fp.readlines():
 					btn =  RoundedButton(
 						text = line[:-1].upper(),
@@ -221,7 +269,10 @@ class MyLayout(Widget):
 			fp.close()
 
 		except FileNotFoundError:
-			os.system("touch ../data/dirs.txt")
+			os.system("touch ../data/dirs_list.txt")
+
+		self.ids.info_.disabled = True
+
 
 	def add_item(self, instance):
 		self.add_popup._open(self)
@@ -237,7 +288,7 @@ class MyLayout(Widget):
 		self.ids.info_.text = ""
 		self.ids.info_.disabled = True
 
-		with open(f"../data/{self.dirname}/files.txt",'r') as fp:
+		with open(f"../data/{self.dirname}/files_list.txt",'r') as fp:
 			for topic in fp.readlines():
 				btn =  RoundedButton(
 					text = topic[:-1].upper(),
